.murdock: use a matrix for kconfig board/tests
As support grows we need a cleaner way of enforcing kconfig testing.
This uses the info-kconfig-boards-supported to run all tests that contain an app.test.config for a supported board.
Groups can still be enforced with the TEST_KCONFIG_ENFORCE_APP_GROUPS.
If a board is in TEST_KCONFIG_BOARDS_AVAILABLE, all supported apps should be built.