gnrc_ipv6_nib: add netif option to include RIO in RA
Add a flag with which we can announce downstream subnet(s) via
Route Information Option(s) in Router Advertisements.

This option should be set on an upstream interface that manages e.g
a subnet of the upstream network.
This prevents other nodes in the upstream network from assuming that
nodes in the downstream network are on-link (since they are in the same
subnet).